Overwrite whitelisted section attributes
Wrote a little function to run whitelisted section attributes
through the jinja2 compiler.
I think the way we actually want to go is to do our transformations
in the `ContentSection`'s `.filter()` method, but for the
purposes of this spike, it'll be enough to just leave the logic
in the view.