Pass in dynamic_list question as a section
Gets us the `one-question-per-page` layout and
puts all the validation messages in the right
place, but it doesn't know how to repopulate
the form with a user's old values.

Sucks, but 😓

Also removes that other stupid view that was
there.  Print lines and inline methods everywhere,
what a mess.  God, who _wrote_ this stupid thing.